+++ packages/api/src/models/Timer.ts
@@ -1,0 +1,19 @@
+/* eslint-disable @typescript-eslint/camelcase */
+import { Nullable } from './helpers'
+
+export default interface Timer {
+	id?: string
+	isRunning?: Nullable<boolean>
+}
+
+/**
+ * Convert a Timer response object
+ *
+ * @param timer Account business object
+ */
+export function transformTimerResponse(timer: any): Timer {
+	return {
+		id: timer.id,
+		isRunning: timer.is_running,
+	}
+}