+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## @freshbooks/api
 
+### Unreleased
+
+- Added Project resource
+
+### 1.2.0
+
+- Added TimeEntry resource
+
 ### 1.1.3
 
 - `like` filters no longer automatically append `_like` to the search key.